Added buffers, fixed some errors
Added buffers for all network I/O streams to improve performance.

I sent and received 99,000 Datapackages (each ~25 bytes) to and from the server (ping-pong) over the internet. The average travel time was 95~105 ms. That is still quite slow, I think. Performance must be further improved. But buffers are already doing a good step in that direction.